\name{matrixMathMisc-package}
\alias{matrixMathMisc-package}
\alias{matrixMathMisc}
\docType{package}
\title{
\packageTitle{matrixMathMisc}
}
\description{
\packageDescription{matrixMathMisc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{matrixMathMisc}
\packageIndices{matrixMathMisc}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{matrixMathMisc}

Maintainer: \packageMaintainer{matrixMathMisc}
}
\references{
~~ Literature or other references for background information ~~
TODO
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ TODO ~~
TODO
}
\examples{
~~ # none ~~s
}
